requirements.txt
langchain_community==0.3.20
langchain_core==0.3.49
langchain_mistralai==0.2.10
langchain_postgres==0.0.13
langchain_text_splitters==0.3.7
langgraph==0.3.21
python-dotenv==1.1.0
